Make git wrapper preserve fugitive env vars
vim-fugitive makes use of these variables so that it can open new splits
when git wants to spawn an editor. However, `flatpak-spawn --host` does
not, by default, forward any environment variables to the host.

Special-case the three variables that fugitive happens to use. It might
be nicer to instead forward all GIT_* to the host. 🤷