Stop forwarding git to host
Git is necessary to use Vim package managers like Vundle, and is not
present in the Platform, so ddcaf8fb096c14242bebf88b8528acc04d6854c0 and
0d23ce708269a216d6e3783df7343dccf6103a7a added this wrapper script which
forwards the command out to the host.

Other advanced text editors like Emacs use the SDK as their runtime. Vim
will follow suit.

forward-to-host is left in place because this may be of use for
special tools available on the host but not the SDK.